requirements.txt
pdfplumber
pandas
numpy
sentence-transformers
faiss-cpu
